requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
pyasn1==0.4.8
pycryptodome==3.15.0
python-decouple==3.6
requests==2.28.0
rsa==4.8
urllib3==1.26.9
